services/gateway/requirements.txt
alpaca-py>=0.32.0
websockets>=12.0
asyncpg>=0.29.0
questdb>=1.2.0
redis>=5.0.0
python-dotenv>=1.0.0
